Fix incorrect conversions_bucket definition in creative excellence
When defining `conversions_bucket` we check whether a particular
campaign optimizes towards a conversion or conversion set that does not
have at least one WEBPAGE conversion action.
To do so we need to filter out campaigns with empty selective
optimization or empty custom conversion actions.

Since a given campaign might have a custom conversion action that
consists of a mix between WEBPAGE and non-WEBPAGE conversion actions we
need to identify campaigns that has ALL conversion actions that are not
WEBPAGE ones.

Change-Id: Ibdf02e2fb25c51b5fd4c0c9af7a1a59f8038c56b